Break some of the internal bits out of pffft.hpp for better readability.
